% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_beta.R
\name{est_beta}
\alias{est_beta}
\title{main function}
\usage{
est_beta(time,delta,vars,wt,model,se.cal)
}
\arguments{
\item{time}{time}

\item{delta}{censoring indicator}

\item{vars}{covariate matrix}

\item{wt}{weight}

\item{model}{specify model type}

\item{se.cal}{choose whether need to calculate standard error}

}
\value{
estimates and standard error of parameters
}
\description{
main function
}
\author{
Peng Jin and Mengling Liu
}
