% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J_mat.R
\name{J_mat}
\alias{J_mat}
\title{Jacobian matrix calculation}
\usage{
J_mat(data_order, z_mat, beta0, model = "exp")
}
\arguments{
\item{data_order}{a ordered dataset by observed time}

\item{z_mat}{covariate matrix}

\item{beta0}{parameter coefficients}

\item{model}{specify additive or proportional MRL model}
}
\value{
jacobian matrix
}
\description{
Jacobian matrix calculation
}
\author{
Peng Jin and Mengling Liu
}
