% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_add.R
\name{score_add}
\alias{score_add}
\title{Score function for additive MRL models}
\usage{
score_add(data_order, z_mat, beta0)
}
\arguments{
\item{data_order}{a ordered dataset by observed time}

\item{z_mat}{covariate matrix}

\item{beta0}{parameter coefficients}
}
\value{
score for each parameter
}
\description{
Score function for additive MRL models
}
\author{
Peng Jin and Mengling Liu
}
