% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_exp.R
\name{score_exp}
\alias{score_exp}
\title{Score function for exponential MRL models}
\usage{
score_exp(data_order, z_mat, beta0)
}
\arguments{
\item{data_order}{dataset ordered by observed time}

\item{z_mat}{covariate matrix}

\item{beta0}{parameter coefficients}
}
\value{
score for each parameter
}
\description{
Score function for exponential MRL models
}
\author{
Peng Jin and Mengling Liu
}
