% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ub_exp.R
\name{ub_exp}
\alias{ub_exp}
\title{Function to calculate derivative}
\usage{
ub_exp(data_order, z_mat, beta0, u, be)
}
\arguments{
\item{data_order}{a ordered dataset by observed time}

\item{z_mat}{covariate matrix}

\item{beta0}{parameter coefficients}

\item{u}{uth number of parameters}

\item{be}{bth number of parameters}
}
\value{
derivative of dudb
}
\description{
Function to calculate derivative
}
\author{
Peng Jin and Mengling Liu
}
