% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpart_add.R
\name{vpart_add}
\alias{vpart_add}
\title{standard error of estimated parameters}
\usage{
vpart_add(data_order, z_mat, beta0)
}
\arguments{
\item{data_order}{a ordered dataset by observed time}

\item{z_mat}{covariate matrix}

\item{beta0}{parameter coefficients}
}
\value{
SEs of estimator
}
\description{
standard error of estimated parameters
}
\author{
Peng Jin and Mengling Liu
}
